requirements.txt
# packaging
pip==21.1.1
setuptools>=26
twine
wheel>=0.29

# data science
numpy==1.18.5
pandas==1.2.0
scikit-learn==0.22
catboost
xgboost>=0.83

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs

# utilities
six>=1.14
joblib
memoized-property
termcolor
pytest
python-gitlab
httplib2>=0.15.0
urllib3==1.25.4
